modules/single_kafka/requirements.txt
Flask==2.2.2
flask-cors==6.0.1
Flask-API==3.1
werkzeug==2.3.0
pymupdf
aspose-words
kafka-python==2.0.2
docx
pikepdf==6.2.0
python-docx==0.8.11
requests==2.28.1
python-dotenv==1.0.0
gunicorn==21.2.0
